+++ crates/cast/src/main.rs
@@ -1,14 +1,15 @@
-use crate::helpers::scarb_utils::parse_scarb_config;
 use crate::starknet_commands::account::Account;
 use crate::starknet_commands::{
     account, call::Call, declare::Declare, deploy::Deploy, invoke::Invoke, multicall::Multicall,
 };
 use anyhow::Result;
 use camino::Utf8PathBuf;
+use cast::helpers::constants::DEFAULT_ACCOUNTS_FILE;
+use cast::helpers::scarb_utils::{parse_scarb_config, CastConfig};
 use cast::{account_file_exists, get_account, get_block_id, get_provider, print_command_result};
 use clap::{Parser, Subcommand};
+use starknet::providers::Provider;
 
-mod helpers;
 mod starknet_commands;
 
 #[derive(Parser)]
@@ -27,21 +28,13 @@ struct Cli {
     #[clap(short = 'u', long = "url")]
     rpc_url: Option<String>,
 
-    /// Network name, one of: testnet, testnet2, mainnet; overrides network from Scarb.toml
-    #[clap(short = 'n', long)]
-    network: Option<String>,
-
     /// Account name to be used for contract declaration; overrides account from Scarb.toml
     #[clap(short = 'a', long)]
     account: Option<String>,
 
     /// Path to the file holding accounts info
-    #[clap(
-        short = 'f',
-        long = "accounts-file",
-        default_value = "~/.starknet_accounts/starknet_open_zeppelin_accounts.json"
-    )]
-    accounts_file_path: Utf8PathBuf,
+    #[clap(short = 'f', long = "accounts-file")]
+    accounts_file_path: Option<Utf8PathBuf>,
 
     /// If passed, values will be displayed as integers, otherwise as hexes
     #[clap(short, long)]
@@ -85,21 +78,17 @@ enum Commands {
 async fn main() -> Result<()> {
     let cli = Cli::parse();
 
-    let accounts_file_path =
-        Utf8PathBuf::from(shellexpand::tilde(&cli.accounts_file_path).to_string());
+    let mut config = parse_scarb_config(&cli.profile, &cli.path_to_scarb_toml)?;
+    update_cast_config(&mut config, &cli);
 
-    let config = parse_scarb_config(&cli.profile, &cli.path_to_scarb_toml)?;
-
-    let rpc_url = cli.rpc_url.unwrap_or(config.rpc_url);
-    let network = cli.network.unwrap_or(config.network);
-    let account = cli.account.unwrap_or(config.account);
-
-    let provider = get_provider(&rpc_url, &network).await?;
+    let provider = get_provider(&config.rpc_url)?;
 
     match cli.command {
         Commands::Declare(declare) => {
-            account_file_exists(&accounts_file_path)?;
-            let mut account = get_account(&account, &accounts_file_path, &provider, &network)?;
+            account_file_exists(&config.accounts_file)?;
+            let chain_id = provider.chain_id().await?;
+            let mut account =
+                get_account(&config.account, &config.accounts_file, &provider, chain_id)?;
 
             let mut result = starknet_commands::declare::declare(
                 &declare.contract,
@@ -115,8 +104,9 @@ async fn main() -> Result<()> {
             Ok(())
         }
         Commands::Deploy(deploy) => {
-            account_file_exists(&accounts_file_path)?;
-            let account = get_account(&account, &accounts_file_path, &provider, &network)?;
+            account_file_exists(&config.accounts_file)?;
+            let chain_id = provider.chain_id().await?;
+            let account = get_account(&config.account, &config.accounts_file, &provider, chain_id)?;
 
             let mut result = starknet_commands::deploy::deploy(
                 deploy.class_hash,
@@ -148,8 +138,10 @@ async fn main() -> Result<()> {
             Ok(())
         }
         Commands::Invoke(invoke) => {
-            account_file_exists(&accounts_file_path)?;
-            let mut account = get_account(&account, &accounts_file_path, &provider, &network)?;
+            account_file_exists(&config.accounts_file)?;
+            let chain_id = provider.chain_id().await?;
+            let mut account =
+                get_account(&config.account, &config.accounts_file, &provider, chain_id)?;
             let mut result = starknet_commands::invoke::invoke(
                 invoke.contract_address,
                 &invoke.function,
@@ -173,9 +165,10 @@ async fn main() -> Result<()> {
                     println!("{result}");
                 }
                 starknet_commands::multicall::Commands::Run(run) => {
-                    account_file_exists(&accounts_file_path)?;
+                    account_file_exists(&config.accounts_file)?;
+                    let chain_id = provider.chain_id().await?;
                     let mut account =
-                        get_account(&account, &accounts_file_path, &provider, &network)?;
+                        get_account(&config.account, &config.accounts_file, &provider, chain_id)?;
                     let mut result = starknet_commands::multicall::run::run(
                         &run.path,
                         &mut account,
@@ -191,13 +184,13 @@ async fn main() -> Result<()> {
         }
         Commands::Account(account) => match account.command {
             account::Commands::Create(create) => {
+                let chain_id = provider.chain_id().await?;
+                config.account = create.name;
                 let mut result = starknet_commands::account::create::create(
+                    &config,
                     &provider,
-                    rpc_url,
-                    accounts_file_path,
                     cli.path_to_scarb_toml,
-                    create.name,
-                    &network,
+                    chain_id,
                     create.salt,
                     create.add_profile,
                     create.class_hash,
@@ -208,12 +201,13 @@ async fn main() -> Result<()> {
                 Ok(())
             }
             account::Commands::Deploy(deploy) => {
-                account_file_exists(&accounts_file_path)?;
+                account_file_exists(&config.accounts_file)?;
+                let chain_id = provider.chain_id().await?;
                 let mut result = starknet_commands::account::deploy::deploy(
                     &provider,
-                    accounts_file_path,
+                    config.accounts_file,
                     deploy.name,
-                    &network,
+                    chain_id,
                     deploy.max_fee,
                     cli.wait,
                     deploy.class_hash,
@@ -225,4 +219,22 @@ async fn main() -> Result<()> {
             }
         },
     }
+}
+
+fn update_cast_config(config: &mut CastConfig, cli: &Cli) {
+    macro_rules! clone_or_else {
+        ($field:expr, $config_field:expr) => {
+            $field.clone().unwrap_or_else(|| $config_field.clone())
+        };
+    }
+
+    config.rpc_url = clone_or_else!(cli.rpc_url, config.rpc_url);
+    config.account = clone_or_else!(cli.account, config.account);
+
+    if config.accounts_file == Utf8PathBuf::default() {
+        config.accounts_file = Utf8PathBuf::from(DEFAULT_ACCOUNTS_FILE);
+    }
+    let new_accounts_file = clone_or_else!(cli.accounts_file_path, config.accounts_file);
+
+    config.accounts_file = Utf8PathBuf::from(shellexpand::tilde(&new_accounts_file).to_string());
 }