+++ crates/cast/src/starknet_commands/account/create.rs
@@ -1,9 +1,9 @@
-use crate::helpers::constants::OZ_CLASS_HASH;
-use crate::helpers::response_structs::AccountCreateResponse;
-use crate::helpers::scarb_utils::{get_property, get_scarb_manifest};
-use anyhow::{anyhow, Context, Result};
+use anyhow::{anyhow, bail, Context, Result};
 use camino::Utf8PathBuf;
-use cast::{extract_or_generate_salt, get_network, parse_number};
+use cast::helpers::constants::OZ_CLASS_HASH;
+use cast::helpers::response_structs::AccountCreateResponse;
+use cast::helpers::scarb_utils::{get_package_tool_sncast, get_scarb_manifest, CastConfig};
+use cast::{chain_id_to_network_name, decode_chain_id, extract_or_generate_salt, parse_number};
 use clap::Args;
 use serde_json::json;
 use starknet::accounts::{AccountFactory, OpenZeppelinAccountFactory};
@@ -38,12 +38,10 @@ pub struct Create {
 
 #[allow(clippy::too_many_arguments)]
 pub async fn create(
+    config: &CastConfig,
     provider: &JsonRpcClient<HttpTransport>,
-    url: String,
-    accounts_file_path: Utf8PathBuf,
     path_to_scarb_toml: Option<Utf8PathBuf>,
-    name: String,
-    network: &str,
+    chain_id: FieldElement,
     maybe_salt: Option<FieldElement>,
     add_profile: bool,
     class_hash: Option<String>,
@@ -68,34 +66,47 @@ pub async fn create(
         let signer = LocalWallet::from_signing_key(private_key.clone());
         let factory = OpenZeppelinAccountFactory::new(
             parse_number(oz_class_hash)?,
-            get_network(network)?.get_chain_id(),
+            chain_id,
             signer,
             provider,
         )
         .await?;
         let deployment = factory.deploy(salt);
 
-        deployment.estimate_fee().await?.overall_fee
+        let fee_estimate = deployment.estimate_fee().await;
+
+        if let Err(err) = &fee_estimate {
+            if err
+                .to_string()
+                .contains("StarknetErrorCode.UNDECLARED_CLASS")
+            {
+                bail!("The class {oz_class_hash} is undeclared, try using --class-hash with a class hash that is already declared");
+            }
+        }
+
+        fee_estimate?.overall_fee
     };
 
-    if !accounts_file_path.exists() {
-        std::fs::create_dir_all(accounts_file_path.clone().parent().unwrap())?;
-        std::fs::write(accounts_file_path.clone(), "{}")?;
+    if !config.accounts_file.exists() {
+        std::fs::create_dir_all(config.accounts_file.clone().parent().unwrap())?;
+        std::fs::write(config.accounts_file.clone(), "{}")?;
     }
 
-    let contents = std::fs::read_to_string(accounts_file_path.clone())?;
+    let contents = std::fs::read_to_string(config.accounts_file.clone())?;
     let mut items: serde_json::Value = serde_json::from_str(&contents)
-        .map_err(|_| anyhow!("Failed to parse accounts file at {accounts_file_path}"))?;
+        .map_err(|_| anyhow!("Failed to parse accounts file at {}", config.accounts_file))?;
 
-    let network_value = get_network(network)?.get_value();
+    let network_name = chain_id_to_network_name(chain_id);
 
-    if !items[network_value][&name].is_null() {
+    if !items[&network_name][&config.account].is_null() {
         return Err(anyhow!(
-            "Account with provided name already exists in this network"
+            "Account with name {} already exists in network with chain_id {}",
+            &config.account,
+            decode_chain_id(chain_id)
         ));
     }
 
-    items[network_value][&name] = json!({
+    items[&network_name][&config.account] = json!({
         "private_key": format!("{:#x}", private_key.secret_scalar()),
         "public_key": format!("{:#x}", public_key.scalar()),
         "address": format!("{address:#x}"),
@@ -104,19 +115,14 @@ pub async fn create(
     });
 
     if add_profile {
-        match add_created_profile_to_configuration(
-            &path_to_scarb_toml,
-            name,
-            network.to_string(),
-            url,
-        ) {
+        match add_created_profile_to_configuration(&path_to_scarb_toml, config) {
             Ok(()) => {}
             Err(err) => return Err(anyhow!(err)),
         };
     }
 
     std::fs::write(
-        accounts_file_path.clone(),
+        config.accounts_file.clone(),
         serde_json::to_string_pretty(&items).unwrap(),
     )?;
 
@@ -145,9 +151,7 @@ pub async fn create(
 
 pub fn add_created_profile_to_configuration(
     path_to_scarb_toml: &Option<Utf8PathBuf>,
-    name: String,
-    network: String,
-    url: String,
+    config: &CastConfig,
 ) -> Result<()> {
     let manifest_path = path_to_scarb_toml.clone().unwrap_or_else(|| {
         get_scarb_manifest().expect("Failed to obtain manifest path from scarb")
@@ -162,24 +166,30 @@ pub fn add_created_profile_to_configuration(
         )
         .unwrap();
 
-    let package = &metadata.packages[0].manifest_metadata.tool;
-
-    let property = get_property(package, &Some(name.clone()), "account");
-    if property.is_ok() {
-        return Err(anyhow!(
-            "Failed to add {name} profile to the Scarb.toml. Profile already exists"
-        ));
+    if let Ok(tool_sncast) = get_package_tool_sncast(&metadata) {
+        let property = tool_sncast
+            .get(&config.account)
+            .and_then(|profile_| profile_.get("account"));
+        if property.is_some() {
+            bail!(
+                "Failed to add {} profile to the Scarb.toml. Profile already exists",
+                config.account
+            );
+        }
     }
 
     let toml_string = {
         let mut tool_sncast = toml::value::Table::new();
         let mut new_profile = toml::value::Table::new();
 
-        new_profile.insert("network".to_string(), Value::String(network));
-        new_profile.insert("url".to_string(), Value::String(url));
-        new_profile.insert("account".to_string(), Value::String(name.clone()));
+        new_profile.insert("url".to_string(), Value::String(config.rpc_url.clone()));
+        new_profile.insert("account".to_string(), Value::String(config.account.clone()));
+        new_profile.insert(
+            "accounts-file".to_string(),
+            Value::String(config.accounts_file.to_string()),
+        );
 
-        tool_sncast.insert(name, Value::Table(new_profile));
+        tool_sncast.insert(config.account.clone(), Value::Table(new_profile));
 
         let mut tool = toml::value::Table::new();
         tool.insert("sncast".to_string(), Value::Table(tool_sncast));
@@ -204,36 +214,38 @@ pub fn add_created_profile_to_configuration(
 #[cfg(test)]
 mod tests {
     use crate::starknet_commands::account::create::add_created_profile_to_configuration;
+    use cast::helpers::constants::DEFAULT_ACCOUNTS_FILE;
+    use cast::helpers::scarb_utils::CastConfig;
     use sealed_test::prelude::rusty_fork_test;
     use sealed_test::prelude::sealed_test;
     use std::fs;
 
     #[sealed_test(files = ["tests/data/contracts/v1/balance/Scarb.toml"])]
     fn test_add_created_profile_to_configuration_happy_case() {
-        let res = add_created_profile_to_configuration(
-            &None,
-            String::from("some-name"),
-            String::from("some-net"),
-            String::from("http://some-url"),
-        );
+        let config = CastConfig {
+            rpc_url: String::from("http://some-url"),
+            account: String::from("some-name"),
+            accounts_file: "accounts".into(),
+        };
+        let res = add_created_profile_to_configuration(&None, &config);
 
         assert!(res.is_ok());
 
         let contents = fs::read_to_string("Scarb.toml").expect("Unable to read Scarb.toml");
         assert!(contents.contains("[tool.sncast.some-name]"));
         assert!(contents.contains("account = \"some-name\""));
-        assert!(contents.contains("network = \"some-net\""));
         assert!(contents.contains("url = \"http://some-url\""));
+        assert!(contents.contains("accounts-file = \"accounts\""));
     }
 
     #[sealed_test(files = ["tests/data/contracts/v1/balance/Scarb.toml"])]
     fn test_add_created_profile_to_configuration_profile_already_exists() {
-        let res = add_created_profile_to_configuration(
-            &None,
-            String::from("myprofile"),
-            String::from("some-net"),
-            String::from("http://some-url"),
-        );
+        let config = CastConfig {
+            rpc_url: String::from("http://some-url"),
+            account: String::from("myprofile"),
+            accounts_file: DEFAULT_ACCOUNTS_FILE.into(),
+        };
+        let res = add_created_profile_to_configuration(&None, &config);
 
         assert!(res.is_err());
     }