+++ crates/forge/src/cheatcodes_hint_processor.rs
@@ -4,7 +4,7 @@ use std::path::PathBuf;
 
 use crate::scarb::StarknetContractArtifacts;
 use anyhow::{anyhow, Result};
-use blockifier::abi::abi_utils::selector_from_name;
+use blockifier::execution::deprecated_syscalls::DeprecatedSyscallSelector;
 use blockifier::execution::execution_utils::{felt_to_stark_felt, stark_felt_to_felt};
 use cairo_felt::Felt252;
 use cairo_vm::hint_processor::hint_processor_definition::HintProcessorLogic;
@@ -21,7 +21,7 @@ use cheatnet::{
 };
 use num_traits::ToPrimitive;
 use serde::Deserialize;
-use starknet_api::core::{ContractAddress, PatriciaKey};
+use starknet_api::core::{ClassHash, ContractAddress, PatriciaKey};
 use starknet_api::hash::StarkFelt;
 
 use cairo_lang_casm::hints::{Hint, StarknetHint};
@@ -106,7 +106,15 @@ impl HintProcessorLogic for CairoHintProcessor<'_> {
             );
         }
         if let Some(Hint::Starknet(StarknetHint::SystemCall { system })) = maybe_extended_hint {
-            return execute_syscall(system, vm, &mut self.cheatnet_state);
+            return execute_syscall(
+                system,
+                vm,
+                &mut self.cheatnet_state,
+                exec_scopes,
+                hint_data,
+                constants,
+                &mut self.original_cairo_hint_processor,
+            );
         }
         self.original_cairo_hint_processor
             .execute_hint(vm, exec_scopes, hint_data, constants)
@@ -231,21 +239,20 @@ impl CairoHintProcessor<'_> {
                     inputs[0].clone().to_be_bytes(),
                 )?)?);
                 let function_name = inputs[1].clone();
-                let function_name = as_cairo_short_string(&function_name).unwrap_or_else(|| {
-                    panic!("Failed to convert {function_name:?} to Cairo short str")
-                });
-                let function_name = selector_from_name(function_name.as_str());
 
                 let ret_data_length = inputs[2]
                     .to_usize()
                     .expect("Missing ret_data len in inputs");
-                let mut ret_data = vec![];
-                for felt in inputs.iter().skip(3).take(ret_data_length) {
-                    ret_data.push(felt_to_stark_felt(&felt.clone()));
-                }
+
+                let ret_data = inputs
+                    .iter()
+                    .skip(3)
+                    .take(ret_data_length)
+                    .cloned()
+                    .collect::<Vec<_>>();
 
                 self.cheatnet_state
-                    .start_mock_call(contract_address, function_name, ret_data);
+                    .start_mock_call(contract_address, &function_name, &ret_data);
                 Ok(())
             }
             "stop_mock_call" => {
@@ -253,13 +260,9 @@ impl CairoHintProcessor<'_> {
                     inputs[0].clone().to_be_bytes(),
                 )?)?);
                 let function_name = inputs[1].clone();
-                let function_name = as_cairo_short_string(&function_name).unwrap_or_else(|| {
-                    panic!("Failed to convert {function_name:?} to Cairo short str")
-                });
-                let function_name = selector_from_name(function_name.as_str());
 
                 self.cheatnet_state
-                    .stop_mock_call(contract_address, function_name);
+                    .stop_mock_call(contract_address, &function_name);
                 Ok(())
             }
             "declare" => {
@@ -292,6 +295,7 @@ impl CairoHintProcessor<'_> {
             }
             "deploy" => {
                 let class_hash = inputs[0].clone();
+                let class_hash = ClassHash(StarkFelt::new(class_hash.to_be_bytes()).unwrap());
 
                 let calldata_length = inputs[1].to_usize().unwrap();
                 let calldata = Vec::from(&inputs[2..(2 + calldata_length)]);
@@ -322,6 +326,7 @@ impl CairoHintProcessor<'_> {
             }
             "precalculate_address" => {
                 let class_hash = inputs[0].clone();
+                let class_hash = ClassHash(StarkFelt::new(class_hash.to_be_bytes()).unwrap());
 
                 let calldata_length = inputs[1].to_usize().unwrap();
                 let calldata = Vec::from(&inputs[2..(2 + calldata_length)]);
@@ -400,22 +405,46 @@ fn execute_syscall(
     system: &ResOperand,
     vm: &mut VirtualMachine,
     cheatnet_state: &mut CheatnetState,
+    exec_scopes: &mut ExecutionScopes,
+    hint_data: &Box<dyn Any>,
+    constants: &HashMap<String, Felt252>,
+    original_cairo_hint_processor: &mut OriginalCairoHintProcessor,
 ) -> Result<(), HintError> {
     let (cell, offset) = extract_buffer(system);
     let system_ptr = get_ptr(vm, cell, &offset)?;
 
-    let mut buffer = MemBuffer::new(vm, system_ptr);
+    // We peek into memory to check the selector
+    let selector = DeprecatedSyscallSelector::try_from(felt_to_stark_felt(
+        &vm.get_integer(system_ptr).unwrap(),
+    ))?;
 
-    let selector = buffer.next_felt252().unwrap().to_bytes_be();
+    return match selector {
+        DeprecatedSyscallSelector::CallContract => {
+            execute_call_contract(MemBuffer::new(vm, system_ptr), cheatnet_state);
+            Ok(())
+        }
+        DeprecatedSyscallSelector::Keccak => {
+            original_cairo_hint_processor.execute_hint(vm, exec_scopes, hint_data, constants)
+        }
+        _ => Err(HintError::CustomHint(Box::from(
+            "starknet syscalls (other than CallContract and Keccak) cannot be used in tests"
+                .to_string(),
+        ))),
+    };
+}
 
-    if std::str::from_utf8(&selector).unwrap() != "CallContract" {
-        return Err(HintError::CustomHint(Box::from(
-            "starknet syscalls cannot be used in tests".to_string(),
-        )));
-    }
-
+fn execute_call_contract(mut buffer: MemBuffer, cheatnet_state: &mut CheatnetState) {
+    let _selector = buffer.next_felt252().unwrap();
     let gas_counter = buffer.next_usize().unwrap();
+
     let contract_address = buffer.next_felt252().unwrap().into_owned();
+    let contract_address = ContractAddress(
+        PatriciaKey::try_from(
+            StarkFelt::new(contract_address.to_be_bytes()).expect("Felt conversion failed"),
+        )
+        .expect("PatriciaKey failed"),
+    );
+
     let entry_point_selector = buffer.next_felt252().unwrap().into_owned();
 
     let calldata = buffer.next_arr().unwrap();
@@ -438,8 +467,6 @@ fn execute_syscall(
     buffer.write(Felt252::from(exit_code)).unwrap();
 
     buffer.write_arr(result.iter()).unwrap();
-
-    Ok(())
 }
 
 fn print(inputs: Vec<Felt252>) {